Add dummy car and make car constructor more flexible
The dummy car is added to the traffic array and is updated and drawn in the
animate function.

The car constructor is made more flexible by adding a controlType parameter
and a maxSpeed parameter.

The controlType parameter can be either "KEYS" or "DUMMY".

The maxSpeed parameter is used to set the maxSpeed of the car.