Add color to cars
This commit adds color to the cars.

The color is randomly generated and is used to create a mask for the car image.

The mask is used to create a colorized version of the car image.

The colorized version is then drawn on top of the original car image.

The colorized version is drawn with the multiply composite operation.

This makes the car image look like it is colored.